alltoowell/HW6/requirements.txt

certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
distlib==0.3.4
filelock==3.6.0
finnhub-python==2.4.9
Flask==2.0.2
google==3.0.0
idna==3.3
itsdangerous==2.0.1
joblib==1.1.0
MarkupSafe==2.0.1
more-itertools==8.12.0

platformdirs==2.5.1
python-dateutil==2.8.2
regex==2022.1.18
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
tqdm==4.62.3
urllib3==1.26.8
virtualenv==20.13.1
Werkzeug==2.0.3
